[proliferate] split session runtime summary and history
